Refactoring tests for MPI
